% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jme4.R
\name{jme4}
\alias{jme4}
\title{jme4}
\usage{
jme4(formula, data, family = NULL)
}
\arguments{
\item{formula}{formula}

\item{data}{data object}

\item{family}{family}
}
\description{
Fit a mixed model using julia
}
